requirements.txt
aws-cdk.core
aws-cdk.aws_ec2
aws-cdk.aws_ecs
aws-cdk.aws_ecr
aws-cdk.aws_ecs_patterns
aws-cdk.aws-route53

# Work around for jsii#413
aws-cdk.aws-autoscaling-common
